+++ src/FuelClient_TEST.cc
@@ -41,7 +41,8 @@ using namespace ignft;
 /// Taken from LocalCache_TEST
 void createLocal1(ClientConfig &_conf)
 {
-  common::createDirectories("test_cache/localhost:8007/alice/models/My Model/meshes");
+  common::createDirectories(
+      "test_cache/localhost:8007/alice/models/My Model/meshes");
 
   {
     std::ofstream fout(
@@ -193,8 +194,8 @@ TEST(FuelClient, ParseModelFileURL)
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://api.ignitionfuel.org/1.0/openrobotics/models/Cordless Drill/\
-files/meshes/cordless_drill.dae"};
+      "https://api.ignitionfuel.org/1.0/openrobotics/models/Cordless Drill/"
+      "files/meshes/cordless_drill.dae"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().URL(), "https://api.ignitionfuel.org");
@@ -213,8 +214,8 @@ files/meshes/cordless_drill.dae"};
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://api.ignitionfuel.org/1.0/openrobotics/models/Pine Tree/\
-files/materials/scripts/pine_tree.material"};
+      "https://api.ignitionfuel.org/1.0/openrobotics/models/Pine Tree/"
+      "files/materials/scripts/pine_tree.material"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().URL(), "https://api.ignitionfuel.org");
@@ -233,8 +234,8 @@ files/materials/scripts/pine_tree.material"};
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://api.ignitionfuel.org/5.0/openrobotics/models/Pine Tree/\
-files/model.sdf"};
+      "https://api.ignitionfuel.org/5.0/openrobotics/models/Pine Tree/"
+      "files/model.sdf"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().URL(), "https://api.ignitionfuel.org");
@@ -250,8 +251,8 @@ files/model.sdf"};
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://api.ignitionfuel.org/openrobotics/models/Pine Tree/\
-files/materials/scripts/pine_tree.material"};
+      "https://api.ignitionfuel.org/openrobotics/models/Pine Tree/"
+      "files/materials/scripts/pine_tree.material"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().URL(), "https://api.ignitionfuel.org");
@@ -270,8 +271,8 @@ files/materials/scripts/pine_tree.material"};
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://api.ignitionfuel.org/openrobotics/models/Pine Tree/\
-files/materials/scripts/pine_tree.material"};
+      "https://api.ignitionfuel.org/openrobotics/models/Pine Tree/"
+      "files/materials/scripts/pine_tree.material"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().URL(), "https://api.ignitionfuel.org");
@@ -441,7 +442,8 @@ TEST(FuelClient, CachedModel)
   // Non-cached model file
   {
     common::URI url{
-      "http://localhost:8007/1.0/alice/models/My Model/files/meshes/banana.dae"};
+      "http://localhost:8007/1.0/alice/models/My Model/files/meshes/banana.dae"
+    };
     std::string path;
     auto result = client.CachedModelFile(url, path);
     EXPECT_FALSE(result);